backend/requirements.txt
aiohttp==3.5.4
python-socketio==4.0.1
aiohttp_session[secure]==2.7.0
ldap3==2.6
aumbry[yaml]==0.10.0
sqlalchemy==1.3
sqlalchemy-aio==0.14.0
alembic==1.0.10